working on pseudo malloc and problems with some flags in mmap on linux

+++ buddy_allocator.c
@@ -112,7 +112,7 @@ void BuddyAllocator_releaseBuddy(BuddyAllocator* alloc, int idx){
 }
 
 //allocates memory
-void* BuddyAllocator_malloc(BuddyAllocator* alloc, int size) {
+void* BuddyAllocator_malloc(BuddyAllocator* alloc, size_t size) {
   // we determine the level of the page
 
   int level = alloc->num_levels -1;
